changed 1 to (long)1 in if statement

+++ src/MemoryIO.java
@@ -16,7 +16,7 @@ public class MemoryIO implements Clockable {
 	public void clockEdge() {
 		if(memRead.getValue().equals((long)1)){
 			readData.setValue(mem.loadWord(address.getValue()));
-		} else if(memWrite.getValue().equals(1)){
+		} else if(memWrite.getValue().equals((long)1)){
 			mem.storeWord(address.getValue(), writeData.getValue());
 		}
 	}
